Extract LocalModpackCard and ModpackEditDialog from Apply
Apply.tsx was 680 lines and is about to absorb two more sections. Move the
per-modpack card and the create/update dialog into their own components so
the fold's diff stays readable. Behaviour is unchanged and Apply.test.tsx is
untouched.

The dialog now owns its edit draft and re-seeds it whenever it opens, which
reproduces the parent's previous set-then-open sequence.

Also drop the dead reference comparison in updateModpacks: it compared a
fresh array against a value captured at mount, so it never short-circuited.
